#5060 Add related id to ezid; consolidate metadata code
